Doesn't work yet but other things on this branch were worth it

+++ plume.h
@@ -6,5 +6,5 @@
 #include "plume_map.h"
 #include "color_wheels.h"
 
-void render_stream_pixels(unsigned int peakToPeak, bool is_beat);
-void render_shoot_pixels(unsigned int peakToPeak, bool is_beat);
+void render_stream_pixels(uint8_t peakToPeak, bool is_beat);
+void render_shoot_pixels(uint8_t peakToPeak, bool is_beat);
